Adding more unit tests to radiation source model (#3333)
* moved some radiation tool function tests from test_radiation_profile to test_radiation_tools

* add test for radiative loss
